12/16/2023 11:55am
Redesigned and waveform and everything else turned into js non react.
All files in the folder called "new" which is inside the original "recorder-demo" folder.

newindex.html is the main html file,
newapp.js is the primary js file that handles most of the scripts,
newrecorder.js is largely the same as originally (I changed what I think is the file reference at the top),
and waveform.js is the js file with the waveform stuff, I expored then imported it into newapp.js for more organised code

The upload thing I'm not sure if it works or not, mukit pls handle it. If the original new version that is running on js works then this should also, but idk how to test it

If you change the file names remember to change the src in newindex.html AND ALSO newrecorder.js (I think the thing at the top of newrecorder.js is a file reference? I'm not 100% sure, heck I don't even understand anything in newrecorder.js)